pyproject.toml
[tool.poetry]
name = "tweet-refiner"
version = "0.1.0"
description = "An application that refines and posts tweets using LLMs"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
tweepy = "^4.14.0"
streamlit = "^1.40.2"
python-dotenv = "^1.0.0"
langchain = "^0.3.9"
langchain-openai = "^0.2.10"

[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
